move query plan execution to another module
